update:exe heredoc

+++ include/execution.h
@@ -6,7 +6,7 @@
 /*   By: tsadouk <tsadouk@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 09:48:47 by cblonde           #+#    #+#             */
-/*   Updated: 2024/07/07 18:29:27 by cblonde          ###   ########.fr       */
+/*   Updated: 2024/07/10 15:06:49 by cblonde          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 # include <sys/wait.h>
 
 bool	ft_execution(t_parse *parse);
-bool	ft_exec_redirect(t_parse *parse, size_t index);
+bool	ft_exec_redirect(t_parse *parse, int index);
 void	ft_get_path(t_parse *parse);
 void	ft_exec(t_parse *parse, t_object *task, size_t i);
 void	ft_exec_builtin(t_parse *parse, t_object *task);
